moved users filling functions to new folder
